transcription/requirements.txt
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2
google-auth-oauthlib
moviepy
python-dotenv
